+++ src/auth/LoginButton.vue
@@ -1,11 +1,20 @@
 <script setup lang="ts">
-import { useRoute } from "vue-router";
+import { useRoute, useRouter } from "vue-router";
 import { PhSignIn } from "@phosphor-icons/vue";
+import { whenever } from "@vueuse/core";
+import { useAuth } from "./auth.composable";
 import { getLoginUrl } from "@/auth/sbAuth";
 import { unarray } from "@/util";
 import UrlButton from "@/components/UrlButton.vue";
 
 const route = useRoute();
+const router = useRouter();
+const { isAuthenticated } = useAuth();
+
+whenever(isAuthenticated, () => {
+  const destination = unarray(route.query.destination);
+  router.push(destination || "/corpus");
+});
 </script>
 
 <template>