+++ src/auth/auth.composable.ts
@@ -1,15 +1,7 @@
-import { computed, ref } from "vue";
-import { useRouter, useRoute } from "vue-router";
+import { computed } from "vue";
 import { useI18n } from "vue-i18n";
-import { retry } from "@/util";
-import api from "@/api/api";
 import useSpin from "@/spin/spin.composable";
-import {
-  fetchJwt,
-  hasAccess,
-  decodeJwt,
-  type JwtSbPayload,
-} from "@/auth/sbAuth";
+import { hasAccess, checkJwt, payload } from "@/auth/sbAuth";
 import useMessenger from "@/message/messenger.composable";
 
 /**
@@ -27,39 +19,20 @@ let jwtPromise: Promise<unknown> | undefined = undefined;
  */
 let refreshTimer: NodeJS.Timeout | undefined = undefined;
 
-const payload = ref<JwtSbPayload>();
-
 export function useAuth() {
-  const router = useRouter();
-  const route = useRoute();
   const { spin, isPending } = useSpin();
   const { alert } = useMessenger();
   const { t } = useI18n();
 
   const isAuthenticated = computed<boolean>(() => !!payload.value);
+  /** Indicates whether a jwt request is currently loading. */
+  const isAuthenticating = computed(() => isPending("jwt"));
   const canUserAdmin = computed<boolean>(
     () =>
       !!payload.value && hasAccess(payload.value, "other", "mink-app", "ADMIN"),
   );
   const canUserWrite = computed(() => isAuthenticated.value);
-  /** Indicates whether a jwt request is currently loading. */
-  const isAuthenticating = computed(() => isPending("jwt"));
-
-  /** If not authenticated, redirect to the login page. */
-  async function requireAuthentication(callback?: () => void) {
-    // First, ensure the jwt has been fetched.
-    if (!payload.value) {
-      await refreshJwt();
-    }
-    // If still no jwt, it means the user hasn't authenticated. Show our login page.
-    if (!payload.value) {
-      // By passing current page as destination param, the user will then be redirected back to where they first attempted to go.
-      router.push(`/login?destination=${route.fullPath}`);
-      return;
-    }
-    // When calling `requireAuthentication`, you can optionally specify what should happen upon success.
-    callback?.();
-  }
+  const userName = computed(() => payload.value?.name || payload.value?.email);
 
   /**
    * Fetch JWT, store it and use it for API client.
@@ -68,16 +41,11 @@ export function useAuth() {
    */
   async function refreshJwt() {
     async function fetchAndStoreJwt() {
-      // Fetch JWT. Occasionally it times out, so try a few times before giving up.
-      const jwtValue = await retry(fetchJwt).catch((error) => {
+      // Fetch JWT.
+      await checkJwt().catch((error) => {
         // On error, show message and treat as not authenticated
         alert(`${t("login.fail")}: ${error?.message || error}`);
-        return undefined;
       });
-      // Store it to make username etc available to GUI.
-      payload.value = jwtValue ? decodeJwt(jwtValue) : undefined;
-      // Register it with the API client.
-      api.setJwt(jwtValue);
 
       // Schedule next request shortly before expiration time.
       if (refreshTimer) clearTimeout(refreshTimer);
@@ -96,10 +64,9 @@ export function useAuth() {
   return {
     isAuthenticating,
     isAuthenticated,
-    payload,
     canUserAdmin,
     canUserWrite,
-    requireAuthentication,
+    userName,
     refreshJwt,
   };
 }