+++ src/library/AdminResourcesView.vue
@@ -4,7 +4,6 @@ import { useRouter } from "vue-router";
 import AdminResourcePreview from "@/library/AdminResourcePreview.vue";
 import LayoutSection from "@/components/LayoutSection.vue";
 import PendingContent from "@/spin/PendingContent.vue";
-import { useAuth } from "@/auth/auth.composable";
 import PageTitle from "@/components/PageTitle.vue";
 import { useResourceStore } from "@/store/resource.store";
 import { isCorpus } from "@/store/resource.types";
@@ -14,12 +13,9 @@ import ActionButton from "@/components/ActionButton.vue";
 
 const router = useRouter();
 const resourceStore = useResourceStore();
-const { requireAuthentication, isAuthenticated } = useAuth();
 const { adminMode } = useAdmin();
 
 const previewToggles = reactive<Record<string, boolean>>({});
-
-requireAuthentication();
 
 watch(adminMode, () => {
   // adminMode is undefined initially. If it resolves to false, go to the normal Library view instead.
@@ -44,7 +40,7 @@ async function load(resourceId: string) {
       {{ $t("admin.resources.help") }}
     </HelpBox>
 
-    <LayoutSection v-if="isAuthenticated">
+    <LayoutSection>
       <PendingContent on="corpora" class="my-4 flex flex-col gap-6">
         <PendingContent
           v-for="(resource, resourceId) of resourceStore.resources"