+++ src/router/metadata.routes.ts
@@ -8,20 +8,28 @@ const metadataRoutes: RouteRecordRaw[] = [
   {
     path: "/library/metadata/new",
     component: CreateMetadata,
-    meta: { title: "metadata.new" },
+    meta: {
+      title: "metadata.new",
+      protected: true,
+    },
   },
   {
     path: "/library/metadata/:resourceId",
     component: MetadataView,
+    meta: { protected: true },
     children: [
       {
         path: "",
         component: MetadataOverview,
+        meta: { protected: true },
       },
       {
         path: "delete",
         component: MetadataDelete,
-        meta: { title: "delete" },
+        meta: {
+          title: "delete",
+          protected: true,
+        },
       },
     ],
   },